refactor: move all test cases in seperate place and added app component test
